app/src/main/java/programacion/kimberly/gym/RegisterActivity.kt
package programacion.kimberly.gym

import android.content.Intent
import android.os.Bundle
import android.widget.Button
import android.widget.EditText
import android.widget.Toast
import androidx.appcompat.app.AppCompatActivity
import com.google.firebase.auth.FirebaseAuth

class RegisterActivity : BaseActivity() {

    lateinit var auth: FirebaseAuth
    private lateinit var editTextEmail: EditText
    private lateinit var editTextPassword: EditText
    private lateinit var buttonNext: Button

    override fun onCreate(savedInstanceState: Bundle?) {
        super.onCreate(savedInstanceState)
        setContentView(R.layout.activity_register)

        // Initialize Firebase Auth
        auth = FirebaseAuth.getInstance()

        // Bind views
        editTextEmail = findViewById(R.id.editTextEmail)
        editTextPassword = findViewById(R.id.editTextPassword)
        buttonNext = findViewById(R.id.buttonNext)

        // Next button click listener
        buttonNext.setOnClickListener {
            createAccount()
        }
    }

    private fun createAccount() {
        val email = editTextEmail.text.toString().trim()
        val password = editTextPassword.text.toString().trim()

        if (email.isEmpty() || password.isEmpty()) {
            Toast.makeText(this, "Email and password cannot be empty", Toast.LENGTH_SHORT).show()
            return
        }

        // Firebase authentication to create a user with email and password
        auth.createUserWithEmailAndPassword(email, password)
            .addOnCompleteListener(this) { task ->
                if (task.isSuccessful) {
                    // Sign in success, update UI with the signed-in user's information
                    val user = auth.currentUser
                    val uid = user?.uid
                    Toast.makeText(baseContext, "Authentication successful", Toast.LENGTH_SHORT).show()

                    // Intent to start NextRegisterActivity
                    val intent = Intent(this, NextRegisterActivity::class.java).apply {
                        // Pass the UID to the next activity if needed
                        putExtra("USER_UID", uid)
                    }
                    startActivity(intent)
                    finish() // Finish this activity
                } else {
                    // If sign in fails, display a message to the user.
                    Toast.makeText(baseContext, "Authentication failed: ${task.exception?.message}", Toast.LENGTH_SHORT).show()
                }
            }
    }
}